Sign install.sh only if checksum doesn't exist.

+++ sync.sh
@@ -9,9 +9,12 @@ S3PREFIX="repos/"
 
 TARGET="s3://${S3BUCKET}/${S3PREFIX}${REPONAME}"
 
-echo "I: Sign install.sh..."
-sha256sum install.sh |tee install.sh.sha256
-gpg -a --sign install.sh.sha256
+if [[ ! -f install.sh.sha256 || ! -f install.sh.sha256.asc ]]
+then
+    echo "I: Sign install.sh..."
+    sha256sum install.sh |tee install.sh.sha256
+    gpg -a --sign install.sh.sha256
+fi
 
 echo "I: Sync config, tools..."
 ${S3CMD} sync install.sh ${TARGET}/
